Adding decode function. Making installable package. Adding more options.
