+++ .github/ci.yml
@@ -1,0 +1,20 @@
+name: ci
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+jobs:
+  pytest:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v1
+        with:
+          python-version: '3.11'
+          architecture: 'x64'
+      - run: pip install -r requirements.txt
+      - run: pytest tests
+      - run: ruff wigners_friend/
