+++ .gitignore
@@ -1,3 +1,6 @@
+# Ruff
+.ruff_cache
+
 # Mac
 *.DS_Store
 # Byte-compiled / optimized / DLL files
