--- experiment.py
@@ -1,43 +1,0 @@
-"""Replicates statistics from https://www.wignersfriends.com/"""
-from toqito.matrices import standard_basis
-import numpy as np
-
-e0, e1 = standard_basis(2, 1)
-e00 = np.kron(e0, e0)
-e01 = np.kron(e0, e1)
-e10 = np.kron(e1, e0)
-e11 = np.kron(e1, e1)
-phi0 = 1/np.sqrt(2) * (e0 + e1)
-phi1 = 1/np.sqrt(2) * (e0 - e1)
-sys = 1/np.sqrt(3) * (np.kron(e0, e0) + np.kron(e0, e1) + np.kron(e1, e1))
-
-def obs(v):
-    return [e0, e1] if v == 1 else [phi0, phi1]
-
-def prob(proj, state):
-    return state.conj().T @ proj @ state
-
-def experiment(a, b, x, y):
-    proj_a = np.outer(obs(x)[a], obs(x)[a])
-    proj_b = np.outer(obs(y)[b], obs(y)[b])
-
-    proj = np.kron(proj_a, proj_b)
-
-    return prob(proj, sys)
-
-
-def phi(x, angles):
-    e0, e1 = standard_basis(2)
-    return 1/np.sqrt(2) * (e0 + np.exp(1j * angles[x]) * e1)
-
-def psi(y, angles, beta):
-    e0, e1 = standard_basis(2)
-    return 1/np.sqrt(2) * (e0 + np.exp(1j * (beta - angles[y])) * e1)
-
-def a(x, angles):
-    return 2 * np.outer(phi(x, angles), phi(x, angles)) - np.outer(e0, e0) - np.outer(e1, e1)
-
-def b(y, angles, beta):
-    return 2 * np.outer(psi(y, angles, beta), psi(y, angles, beta)) - np.outer(e0, e0) - np.outer(e1, e1)
-
-print(experiment(1,1,2,2))
