+++ requirements.txt
@@ -1,7 +1,6 @@
-toqito
-qiskit
-cirq
-pylatexenc
-qiskit-aer
-python-dotenv
-qiskit_ibm_runtime
+qiskit==0.44.2
+qiskit-aer==0.12.2
+python-dotenv==1.0.0
+qiskit-ibm-runtime==0.16.1
+numpy==1.23.5
+ruff==0.1.7
