--- src/main.py
@@ -1,43 +1,0 @@
-import itertools
-import qiskit
-from qiskit.providers import Backend
-from qiskit_aer.noise import NoiseModel
-
-from observer import Observer
-from ewfs_circuit import ewfs
-from violations import compute_inequalities
-from config import (SETTINGS, ANGLES, BETA, BACKEND, NOISE_MODEL, SHOTS)
-
-
-def generate_all_experiments(
-    backend: Backend,
-    noise_model: NoiseModel,
-    shots: float,
-    angles: list[float],
-    beta: float
-) -> dict[tuple[Observer, Observer], list[float]]:
-    """Generate probabilitites for all combinations of experimental settings."""
-    all_experiment_combos = list(itertools.product(SETTINGS, repeat=2))
-    
-    results = {}
-    for alice, bob in all_experiment_combos:
-        ewfs_circuit = ewfs(alice, bob, angles, beta)
-
-        job = qiskit.execute(
-            experiments=ewfs_circuit,
-            backend=backend,
-            noise_model=noise_model,
-            basis_gates=noise_model.basis_gates if noise_model is not None else None,
-            shots=shots,
-        )
-        counts = job.result().get_counts()
-        
-        # Convert counts to probabilities.
-        probabilities = {key[::-1]: value / shots for key, value in counts.items()}
-
-        results[(alice, bob)] = probabilities
-    return results
-
-
-results = generate_all_experiments(backend=BACKEND, noise_model=NOISE_MODEL, shots=SHOTS, angles=ANGLES, beta=BETA)
-compute_inequalities(results, verbose=True)
