+++ wigners_friend/observer.py
