+++ wigners_friend/setting.py
