+++ wigners_friend/stats.py
@@ -1,0 +1,123 @@
+from wigners_friend.config import (
+    ALICE, BOB,
+    PEEK, REVERSE_1, REVERSE_2,
+    SETTINGS,
+)
+from wigners_friend.setting import Setting
+from wigners_friend.observer import Observer
+
+
+def single_expect(observer: Observer, setting: Setting, results: dict):
+    """Compute single expectation values for either Alice or Bob."""
+    if observer is ALICE:
+        ret = 0
+        for settings in results.keys():
+            if settings[ALICE] is setting:
+                probs = results[settings]
+                # <A> = P(00) + P(01) - P(10) - P(11)
+                ret += (
+                    probs.get("00", 0)
+                    + probs.get("01", 0)
+                    - probs.get("10", 0)
+                    - probs.get("11", 0)
+                )
+        return ret / len(SETTINGS)
+    else:
+        ret = 0
+        for settings in results.keys():
+            if settings[BOB] is setting:
+                probs = results[settings]
+                # <B> = P(00) - P(01) + P(10) - P(11)
+                ret += (
+                    probs.get("00", 0)
+                    - probs.get("01", 0)
+                    + probs.get("10", 0)
+                    - probs.get("11", 0)
+                )
+        return ret / len(SETTINGS)
+
+
+def double_expect(settings: list[Setting], results: dict) -> float:
+    """Expectation value of product of two operators."""
+    probs = results[settings]
+    # <AB> = P(00) - P(01) - P(10) + P(11)
+    return (
+        probs.get("00", 0) 
+        - probs.get("01", 0) 
+        - probs.get("10", 0) 
+        + probs.get("11", 0)
+    )    
+
+
+def compute_inequalities(results: dict, verbose: bool = False):
+    A1 = single_expect(ALICE, PEEK, results)
+    B1 = single_expect(BOB, PEEK, results)
+
+    A2 = single_expect(ALICE, REVERSE_1, results)
+    B2 = single_expect(BOB, REVERSE_1, results)
+
+    A3 = single_expect(ALICE, REVERSE_2, results)
+    B3 = single_expect(BOB, REVERSE_2, results)
+
+    A1B1 = double_expect((PEEK, PEEK), results)
+    A1B2 = double_expect((PEEK, REVERSE_1), results)
+    A1B3 = double_expect((PEEK, REVERSE_2), results)
+
+    A2B1 = double_expect((REVERSE_1, PEEK), results)
+    A2B2 = double_expect((REVERSE_1, REVERSE_1), results)
+    A2B3 = double_expect((REVERSE_1, REVERSE_2), results)
+
+    A3B1 = double_expect((REVERSE_2, PEEK), results)
+    A3B2 = double_expect((REVERSE_2, REVERSE_1), results)
+    A3B3 = double_expect((REVERSE_2, REVERSE_2), results)
+    
+    # Local-friendliness inequalities:
+    # Eq. (13) from [1].
+    lf = -A1 - A2 - B1 - B2 - A1B1 - 2*A1B2 - 2*A2B1 + 2*A2B2 - A2B3 - A3B2 - A3B3 - 6
+    # Eq. (15) from [1].
+    I3322 = -A1 + A2 + B1 - B2 + A1B1 - A1B2 - A1B3 - A2B1 + A2B2 - A2B3 - A3B1 - A3B2 - 4 
+
+    # Brukner inequalities:
+    # Eq. (17) from [1].
+    brukner = A1B1 - A1B3 - A2B1 - A2B3 - 2
+    # Eq. (18) from [1].
+    semi_brukner = -A1B2 + A1B3 - A3B2 - A3B3 - 2
+
+    # Bell non-LF 
+    # Eq. (22) from [1].
+    bell_non_lf = A2B2 - A2B3 - A3B2 - A3B3 - 2
+    
+    if verbose:
+        print("******Inequalities******")
+        print(f"{semi_brukner=} -- is violated: {semi_brukner > 0}")
+        print(f"{brukner=} -- is violated: {brukner > 0}")
+        print(f"{lf=} -- is violated: {lf > 0}")
+        print(f"{I3322=} -- is_violated: {I3322 > 0}")
+        print(f"{bell_non_lf=} -- is_violated: {bell_non_lf > 0}")
+        print("**************************")
+        
+        print("******Expectation values******")
+        print(f"{A1=}")
+        print(f"{A2=}")
+        print(f"{A3=}")
+        print(f"{B1=}")
+        print(f"{B2=}")
+        print(f"{B3=}")
+        print(f"{A1B1=}")
+        print(f"{A1B2=}")
+        print(f"{A1B3=}")
+        print(f"{A2B1=}")
+        print(f"{A2B2=}")
+        print(f"{A2B3=}")
+        print(f"{A3B1=}")
+        print(f"{A3B2=}")
+        print(f"{A3B3=}")
+        print("******************************")
+
+    return {
+        "lf": lf,
+        "I3322": I3322,
+        "brukner": brukner,
+        "semi-brukner": semi_brukner,
+        "bell_non_lf": bell_non_lf,
+    }
